requirements.txt
fastapi
uvicorn
jinja2
openai
pandas
python-docx
